# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  encryption.RotateEncryptionPhraseRequest:
    properties:
      phrase:
        type: string
    type: object
...
  notification.NotificationConfigRequest:
    properties:
      channel_config:
        additionalProperties: {}
        type: object
      type:
        type: string
    type: object
  notification.NotificationConfigResponse:
    properties:
      channel_config:
        additionalProperties: {}
        type: object
      type:
        type: string
    type: object
  notification.NotificationTypeResponse:
    properties:
      label:
        type: string
      type:
        type: string
    type: object
  oidc.OidcConfigRequest:
    properties:
      app_url:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      issuer_url:
        type: string
    type: object
  oidc.OidcConfigResponse:
    properties:
      app_url:
        type: string
      client_id:
        type: string
      issuer_url:
        type: string
    type: object
...
info:
  contact: {}
  description: 'All endpoints return responses wrapped in a standard envelope: {"success":
    bool, "data": (schema shown per endpoint), "error": string}. All protected endpoints
    can return 401.'
  title: ez2boot API
  version: "1.0"
paths:
...
  /auth/oidc:
    delete:
      description: '[Admin] delete an OIDC config. No payload or selector, as there
        can be only one configuration.'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      security:
      - BasicAuth: []
      summary: Delete OIDC config
      tags:
      - oidc
    post:
      consumes:
      - application/json
      description: '[Admin] set an OIDC config'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oidc.OidcConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Set OIDC config
      tags:
      - oidc
    put:
      description: '[Admin] get the current OIDC config'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oidc.OidcConfigResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Get OIDC config
      tags:
      - oidc
  /auth/oidc/test:
    post:
      description: '[Admin] test the OIDC connection'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Test OIDC connection
      tags:
      - oidc
  /encryption/passphrase:
    put:
      consumes:
      - application/json
      description: '[Admin] trigger an automated decrypt and re-encrypt of sensitive
        data fields within the database with a supplied phrase. Requires updating
        env var after.'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/encryption.RotateEncryptionPhraseRequest'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Rotate encryption phrase
      tags:
      - encryption
  /notification/types:
    get:
      description: Get all supported notification types
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.NotificationTypeResponse'
            type: array
      security:
      - BasicAuth: []
      summary: Get notification types
      tags:
      - notification
...
  /user/notification:
    delete:
      description: Delete own notification config
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      security:
      - BasicAuth: []
      summary: Delete notification config
      tags:
      - notification
    get:
      description: Get own notification config
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.NotificationConfigResponse'
      security:
      - BasicAuth: []
      summary: Get notification config
      tags:
      - notification
    post:
      consumes:
      - application/json
      description: Set own notification config
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notification.NotificationConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Set notification config
      tags:
      - notification
  /user/notification/test:
    post:
      description: Test own notification config and receive a message
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      security:
      - BasicAuth: []
      summary: Test notification config
      tags:
      - notification
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"